+++ src/Accounts/Accounts/CommonModule/ContextAdapter.cs
@@ -28,6 +28,9 @@ using System.Linq;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.Profile.Properties;
 using Azure.Identity;
+using Microsoft.Azure.Commands.Common.Authentication.Abstractions.Interfaces;
+using Microsoft.Azure.Commands.Common.Authentication.Abstractions.Extensions;
+using Microsoft.Azure.Commands.Common.Authentication.Factories;
 
 namespace Microsoft.Azure.Commands.Common
 {
@@ -75,7 +78,7 @@ namespace Microsoft.Azure.Commands.Common
         {
             prependStep(UniqueId.Instance.SendAsync);
             appendStep(new UserAgent(invocationInfo).SendAsync);
-            appendStep(this.SendHandler(GetDefaultContext(_provider, invocationInfo), AzureEnvironment.Endpoint.ActiveDirectoryServiceEndpointResourceId));
+            appendStep(this.SendHandler(GetDefaultContext(_provider, invocationInfo), AzureEnvironment.Endpoint.ActiveDirectoryServiceEndpointResourceId, new AzureCmdletContext(correlationId)));
         }
 
         internal void AddRequestUserAgentHandler(
@@ -118,7 +121,12 @@ namespace Microsoft.Azure.Commands.Common
                 {
                     endpointResourceIdKey = endpointResourceIdKey ?? AzureEnvironment.Endpoint.ActiveDirectoryServiceEndpointResourceId;
                     var context = GetDefaultContext(_provider, invocationInfo);
-                    return await AuthenticationHelper(context, endpointResourceIdKey, endpointSuffixKey, request, cancelToken, cancelAction, signal, next);
+                    ICmdletContext cmdletContext = AzureCmdletContext.CmdletNone;
+                    if (extensibleParameters.ContainsKey(AuthenticationFactory.CmdletContextParameterName))
+                    {
+                        cmdletContext = extensibleParameters[AuthenticationFactory.CmdletContextParameterName] as ICmdletContext;
+                    }
+                    return await AuthenticationHelper(context, endpointResourceIdKey, endpointSuffixKey, request, cancelToken, cancelAction, signal, next, cmdletContext);
                 });
         }
 
@@ -193,9 +201,10 @@ namespace Microsoft.Azure.Commands.Common
             return string.Empty;
         }
 
-        internal async Task<HttpResponseMessage> AuthenticationHelper(IAzureContext context, string endpointResourceIdKey, string endpointSuffixKey, HttpRequestMessage request, CancellationToken cancelToken, Action cancelAction, SignalDelegate signal, NextDelegate next, TokenAudienceConverterDelegate tokenAudienceConverter = null)
+        internal async Task<HttpResponseMessage> AuthenticationHelper(IAzureContext context, string endpointResourceIdKey, string endpointSuffixKey, HttpRequestMessage request, CancellationToken cancelToken, Action cancelAction, SignalDelegate signal, NextDelegate next, ICmdletContext cmdletContext, TokenAudienceConverterDelegate tokenAudienceConverter = null)
         {
-            IAccessToken accessToken = await AuthorizeRequest(context, request, cancelToken, endpointResourceIdKey, endpointSuffixKey, tokenAudienceConverter);
+            var extensiableParameters = cmdletContext?.ToExtensibleParameters();
+            IAccessToken accessToken = await AuthorizeRequest(context, request, cancelToken, endpointResourceIdKey, endpointSuffixKey, tokenAudienceConverter, extensiableParameters);
             using (var newRequest = await request.CloneWithContent(request.RequestUri, request.Method))
             {
                 var response = await next(request, cancelToken, cancelAction, signal);
@@ -232,13 +241,14 @@ namespace Microsoft.Azure.Commands.Common
         /// </summary>
         /// <param name="context"></param>
         /// <param name="resourceId"></param>
+        /// <param name="cmdletContext"></param>
         /// <returns></returns>
-        internal Func<HttpRequestMessage, CancellationToken, Action, SignalDelegate, NextDelegate, Task<HttpResponseMessage>> SendHandler(IAzureContext context, string resourceId)
+        internal Func<HttpRequestMessage, CancellationToken, Action, SignalDelegate, NextDelegate, Task<HttpResponseMessage>> SendHandler(IAzureContext context, string resourceId, ICmdletContext cmdletContext)
         {
             return async (request, cancelToken, cancelAction, signal, next) =>
             {
                 PatchRequestUri(context, request);
-                return await AuthenticationHelper(context, resourceId, resourceId, request, cancelToken, cancelAction, signal, next);
+                return await AuthenticationHelper(context, resourceId, resourceId, request, cancelToken, cancelAction, signal, next, cmdletContext);
             };
         }
 
@@ -249,12 +259,12 @@ namespace Microsoft.Azure.Commands.Common
         /// <param name="endpointResourceIdKey"></param>
         /// <param name="request"></param>
         /// <param name="endpointSuffixKey"></param>
-        /// <param name="extensibleParamters"></param>
+        /// <param name="extensibleParameters"></param>
         /// <param name="tokenAudienceConverter"></param>
         /// <param name="cancellationToken"></param>
         /// <returns></returns>
         internal async Task<IAccessToken> AuthorizeRequest(IAzureContext context, HttpRequestMessage request, CancellationToken cancellationToken, string endpointResourceIdKey,
-                        string endpointSuffixKey, TokenAudienceConverterDelegate tokenAudienceConverter = null, IDictionary<string, object> extensibleParamters = null)
+                        string endpointSuffixKey, TokenAudienceConverterDelegate tokenAudienceConverter = null, IDictionary<string, object> extensibleParameters = null)
         {
             if (context == null || context.Account == null || context.Environment == null)
             {
@@ -269,7 +279,12 @@ namespace Microsoft.Azure.Commands.Common
                     var tokenAudience = tokenAudienceConverter.Invoke(info.CurEnvEndpointResourceId, info.CurEnvEndpointSuffix, info.BaseEnvEndpointResourceId, info.BaseEnvEndpointSuffix, request.RequestUri);
                     endpointResourceIdKey = tokenAudience ?? endpointResourceIdKey;
                 }
-                var authToken = _authenticator.Authenticate(context.Account, context.Environment, context.Tenant.Id, null, "Never", null, endpointResourceIdKey);
+                var optionalParameters = new Dictionary<string, object>() { { AuthenticationFactory.ResourceIdParameterName, endpointResourceIdKey } };
+                if (extensibleParameters != null && extensibleParameters.ContainsKey(AuthenticationFactory.CmdletContextParameterName))
+                {
+                    optionalParameters.Add(AuthenticationFactory.CmdletContextParameterName, extensibleParameters[AuthenticationFactory.CmdletContextParameterName]);
+                }
+                var authToken = _authenticator.Authenticate(context.Account, context.Environment, context.Tenant.Id, null, "Never", null, optionalParameters);
                 authToken.AuthorizeRequest((type, token) => request.Headers.Authorization = new System.Net.Http.Headers.AuthenticationHeaderValue(type, token));
                 return authToken;
             }, cancellationToken);